+++ src/death.cpp
@@ -29,8 +29,8 @@ static char *center_string(char *centered_str, const char *in_str) {
 void showScoresScreen() {
     char string[100];
 
-    if ((highscore_fp = fopen(MORIA_TOP, "rb")) == NULL) {
-        sprintf(string, "Error opening score file \"%s\"\n", MORIA_TOP);
+    if ((highscore_fp = fopen(MORIA_SCORES, "rb")) == NULL) {
+        sprintf(string, "Error opening score file \"%s\"\n", MORIA_SCORES);
         printMessage(string);
         printMessage(CNIL);
         return;
@@ -261,10 +261,10 @@ static void highscores() {
     }
     (void) strcpy(new_entry.died_from, tmp);
 
-    if ((highscore_fp = fopen(MORIA_TOP, "rb+")) == NULL) {
+    if ((highscore_fp = fopen(MORIA_SCORES, "rb+")) == NULL) {
         char string[100];
 
-        (void) sprintf(string, "Error opening score file \"%s\"\n", MORIA_TOP);
+        (void) sprintf(string, "Error opening score file \"%s\"\n", MORIA_SCORES);
         printMessage(string);
         printMessage(CNIL);
         return;